refactor(kbucket): split kbucket and ktree to separate modules
