+++ src/toxcore/dht/mod.rs
@@ -3,6 +3,7 @@
 
 pub mod packet;
 pub mod kbucket;
+pub mod ktree;
 pub mod packed_node;
 pub mod codec;
 pub mod server;
